+++ R/measure-path.R
@@ -11,7 +11,7 @@ measurePathDependencies <- function() {
 #' @examples
 #' \donttest{
 #' geoJson <- readr::read_file(
-#'   "https://rawgit.com/benbalter/dc-maps/master/maps/ward-2012.geojson"
+#'   "https://raw.githubusercontent.com/benbalter/dc-maps/master/maps/ward-2012.geojson"
 #' )
 #'
 #' leaflet() %>%