+++ R/webGLHeatmap.R
@@ -103,7 +103,7 @@ addWebGLHeatmap <- function(
 #' ## addWebGLGeoJSONHeatmap
 #' \donttest{
 #' geoJson <- readr::read_file(
-#'   "https://rawgit.com/benbalter/dc-maps/master/maps/historic-landmarks-points.geojson"
+#'   "https://raw.githubusercontent.com/benbalter/dc-maps/master/maps/historic-landmarks-points.geojson"
 #' )
 #'
 #' leaflet() %>%